Added "Path-Section-Date"
Whenever a member checks for broken links, grammar, etc… he MUST write
in this file tha Path of the analyzed file(hardware/amd.html or
sowtware/whatever.html), the section of the file (Software, Hardware,
etc..) and the date:
The date must be written in this format: [yyyy/mm/dd]
If a precise date is not found, then approximate or invent a plausible
date for that event / product.
This is important for the Lucas’ script to work.